# sih-backend/contracts/openapi.yaml
openapi: "3.0.3"
info:
  title: Unigraph — Real-Time Crypto Fraud Attribution API
  description: |
    Backend API for Unigraph (SIH26183 — MHA/I4C).

    **Auth schemes:**
    - `bearerAuth` (JWT): all `/api/v1/...` routes except `/api/v1/auth/login` and `/api/v1/auth/refresh`
    - `apiKeyAuth` (X-API-Key header): `/check-wallet` only — used by external VASPs, never by the frontend JWT client

    **Error envelope** (all failures, every endpoint):
    ```json
    { "error": { "code": "string", "message": "string", "details": {} } }
    ```

    **Enums** are defined in `contracts/entities.md` and must never be extended inline.
  version: "0.1.0"
  contact:
    name: Backend Owner (Rehan)
servers:
  - url: http://localhost:8000
    description: Local development

# ── Security schemes ─────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

  # ── Shared schemas ──────────────────────────────────────────────────────────
  schemas:
...
    AlertAction:
      type: string
      enum: [allow, hold, block]
      description: "Closed enum — returned by /check-wallet and stored on Alert records."
    Chain:
      type: string
      enum: [BTC, ETH, TRON, BSC]
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: WALLET_NOT_FOUND
            message:
              type: string
              example: "No wallet found for that address/chain pair."
            details:
              type: object
              additionalProperties: true
              example: { "address": "1A1zP1...", "chain": "BTC" }

    # ── Pagination wrapper ────────────────────────────────────────────────────
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "investigator@i4c.gov.in"
        password:
          type: string
          format: password
          example: "s3cr3t"
    LoginResponse:
      type: object
      required: [access_token, refresh_token, role]
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        role:
          type: string
          enum: [admin, investigator, compliance_viewer]
          example: "investigator"
...
    CheckWalletRequest:
      type: object
      required: [address, chain, amount]
      properties:
        address:
          type: string
          example: "1A1zP1eP5QGefi2DMPTfTL5SLmv7Divf2"
        chain:
          $ref: "#/components/schemas/Chain"
        amount:
          type: number
          format: float
          example: 0.5
    CheckWalletResponse:
      type: object
      required: [risk_score, action]
      properties:
        risk_score:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          example: 0.92
        action:
          $ref: "#/components/schemas/AlertAction"
        case_ref:
          type: string
          nullable: true
          example: "NCRP-2026-009123"
...
  responses:
...
    UnprocessableEntity:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"

# ── Global default auth (JWT bearer) ─────────────────────────────────────────
security:
  - bearerAuth: []

# ── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /api/v1/auth/login:
    post:
      summary: Obtain JWT access + refresh tokens
      operationId: auth_login
      security: [] # no auth — this IS the login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Bad credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: INVALID_CREDENTIALS
                  message: "Email or password is incorrect."
                  details: {}
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
...
  /check-wallet:
    post:
      summary: Real-time risk check for VASP deposit chokepoint (p95 < 200ms)
      description: |
        **Auth: X-API-Key header (apiKeyAuth) — NOT a JWT Bearer token.**
        This endpoint is consumed by external VASP deposit systems, not by the
        React frontend. The frontend client must never call this route.

        Hot path: reads Redis only (`risk:{chain}:{address}`). 
        Never touches Postgres or Neo4j synchronously.
      operationId: check_wallet
      tags: [vasp]
      security:
        - apiKeyAuth: [] # explicitly NOT bearerAuth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckWalletRequest"
      responses:
        "200":
          description: Risk decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckWalletResponse"
              example:
                risk_score: 0.92
                action: "hold"
                case_ref: "NCRP-2026-009123"
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: INVALID_API_KEY
                  message: "The provided X-API-Key is not recognised."
                  details: {}
        "422":
          $ref: "#/components/responses/UnprocessableEntity"

  # ── Cases ──────────────────────────────────────────────────────────────────
...